make it all strict
