+++ src/index.ts
@@ -3,4 +3,4 @@ export * from './errors';
 export * from './reporter';
 export * from './types';
 export * from './validators';
-export * from './strictEnv';
+export * from './freezeObject';
